Refactor: Moved tests to test_webcrawler.py and cleaned up main.py
